fix location for temporary download and Data Availability in package

+++ R/update_ComtradeDA.R
@@ -13,13 +13,16 @@
 #' @keywords Comtrade data availability
 #' @export
 #' @import dplyr comtradr tibble readr rlang
-update_ComtradeDA <- function(directory = "data",
+update_ComtradeDA <- function(directory = system.file('data', package = 'comtradeRggregator'),
                               file = paste0("Comtrade_DataAvailability-", Sys.Date())) {
   file_exists <- list.files(directory, pattern = file)
 
   if (identical(file_exists, character(0))) {
-    file_old <- list.files(directory)
-    file.remove(paste0(directory, "/", file_old))
+    file_old <- list.files(directory, pattern = 'Comtrade_DataAvailability-')
+
+    if (!identical(file_exists, character(0))) {
+      file.remove(paste0(directory, "/", file_old))
+    }
 
     url <- "https://comtrade.un.org/api/refs/da/view?fmt=csv"
     Comtrade_DA <- readr::read_csv(url, col_types = cols())
